+++ src/main/java/ru/practicum/constants/BaseConstants.java
