+++ src/main/java/ru/practicum/constants/Browser.java
