+++ src/main/java/ru/practicum/constants/UserConstants.java
