+++ src/main/java/ru/practicum/pojos/BasicResponse.java
