+++ src/main/java/ru/practicum/pojos/NestedUserInfo.java
