+++ src/main/java/ru/practicum/pojos/SignInRequest.java
