+++ src/main/java/ru/practicum/pojos/SuccessSignInSignUpResponse.java
