+++ src/main/java/ru/practicum/pojos/UserRequest.java
