+++ src/test/java/LogoutTest.java
@@ -1,5 +1,3 @@
-package ru.practicum;
-
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.RestAssured;
 import org.junit.After;
@@ -9,11 +7,11 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.openqa.selenium.WebDriver;
+import ru.practicum.page_objects.LoginPage;
+import ru.practicum.page_objects.MainPage;
 import ru.practicum.api_steps.UsersSteps;
 import ru.practicum.constants.Browser;
 import ru.practicum.page_objects.AccountPage;
-import ru.practicum.page_objects.LoginPage;
-import ru.practicum.page_objects.MainPage;
 import ru.practicum.pojos.SignInRequest;
 import ru.practicum.pojos.SuccessSignInSignUpResponse;
 import ru.practicum.pojos.UserRequest;
@@ -24,15 +22,17 @@ import ru.practicum.utils.UsersUtils;
 import java.time.Duration;
 
 @RunWith(Parameterized.class)
-public class GoToAccountTest {
+public class LogoutTest {
     WebDriver driver;
     MainPage mainPage;
     LoginPage loginPage;
     AccountPage accountPage;
     Browser browserEnum;
-    ConfigFileReader configFileReader = new ConfigFileReader();
+    UserRequest testUser;
+    String accessToken;
+    SignInRequest signInRequest;
 
-    public GoToAccountTest(Browser browserEnum) {
+    public LogoutTest(Browser browserEnum) {
         this.browserEnum = browserEnum;
     }
 
@@ -46,11 +46,22 @@ public class GoToAccountTest {
 
     @Before
     public void init() {
+        testUser = UsersUtils.getUniqueUser();
+        RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
+
+        SuccessSignInSignUpResponse signUpResponse = UsersSteps.createUniqueUser(testUser)
+                .then()
+                .statusCode(200)
+                .extract()
+                .as(SuccessSignInSignUpResponse.class);
+        accessToken = signUpResponse.getAccessToken();
+        signInRequest = new SignInRequest(testUser.getEmail(), testUser.getPassword());
+
         driver = DriverInitializer.getDriver(browserEnum);
-        driver.get(configFileReader.getApplicationUrl());
         mainPage = new MainPage(driver);
-        accountPage = new AccountPage(driver);
         loginPage = new LoginPage(driver);
+        accountPage = new AccountPage(driver);
+        driver.get(new ConfigFileReader().getApplicationUrl() + "/login");
         driver.manage().window().maximize();
         driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
     }
@@ -61,25 +72,14 @@ public class GoToAccountTest {
     }
 
     @Test
-    @DisplayName("Успешный переход по клику на Личный кабинет")
-    public void goToAccountWithAccountButtonSuccess() {
-        UserRequest user = UsersUtils.getUniqueUser();
-        RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
-
-        String accessToken = UsersSteps.createUniqueUser(user)
-                .then()
-                .statusCode(200)
-                .extract()
-                .as(SuccessSignInSignUpResponse.class)
-                .getAccessToken();
-
+    @DisplayName("Выход по кнопке Выйти в личном кабинете")
+    public void logoutWithLogoutButtonSuccess() {
+        loginPage.loginWithCredentials(signInRequest);
         mainPage.clickAccountButton();
-        loginPage.loginWithCredentials(new SignInRequest(user.getEmail(), user.getPassword()));
+        accountPage.clickLogoutButton();
         mainPage.clickAccountButton();
 
-        boolean displayed = accountPage.getProfileButton().isDisplayed();
-        Assert.assertTrue("Личный кабинет не открыт", displayed);
-
-        UsersSteps.deleteUser(accessToken);
+        boolean displayed = loginPage.getSignInButton().isDisplayed();
+        Assert.assertTrue("Выход из личного кабинета не выполнен", displayed);
     }
 }
